requirements.txt
# filepath: c:\Users\Husni kusuma\Desktop\quantum_trading_system\requirements.txt
# Qiskit stack (versi stabil sebelum symengine)
qiskit==0.46.0
qiskit-optimization==0.6.0
qiskit-aer==0.13.3

# Data & Analisis
pandas
numpy
scikit-learn

# API & Utilitas
requests
pyyaml  # For YAML configuration files
redis   # For caching system

# Visualisasi
matplotlib

# Deep Learning
tensorflow

# Data Validation & Quality
jsonschema  # For data schema validation
pydantic    # For data validation and settings management

# Async & Concurrency
aiohttp     # For async HTTP requests
asyncio-throttle  # For rate limiting

# Monitoring & Logging
prometheus-client  # For metrics collection
structlog   # For structured logging

# Development & Testing
pytest      # For testing
pytest-asyncio  # For async testing
black       # Code formatting
flake8      # Code linting
